pyt-image-cls-api/requirements.txt
seldon-core==1.0.0 #一定要是這個版本
numpy==1.17.2
Pillow==6.2.0
torch==1.4
torchvision==0.5.0
